elastic/requirements.txt
elasticsearch
PyMuPDF
python-frontmatter
python-dotenv
fastapi
uvicorn
pyzmq